Added driver registration

+++ NetfilterInstaller/DriverInstaller.cs
@@ -1,7 +1,10 @@
-﻿using System;
+﻿using Microsoft.Win32;
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Runtime.InteropServices;
+using System.Windows.Forms;
 
 namespace NetfilterInstaller
 {
@@ -13,6 +16,64 @@ namespace NetfilterInstaller
         {
             TDI = 100,
             WFP = 200
+        }
+
+        enum RegAction
+        {
+            Register = 300,
+            Unregister = 400
+        }
+
+        #region kernel32.dll methods
+        [DllImport("kernel32.dll", SetLastError = true)]
+        static extern bool Wow64DisableWow64FsRedirection(ref IntPtr ptr);
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        static extern bool Wow64RevertWow64FsRedirection(IntPtr ptr);
+        #endregion
+
+        static void SetUninstallKeyStarted()
+        {
+            RegistryKey netfilterRegKey = Registry.LocalMachine.CreateSubKey("SOFTWARE\\Netfilter\\Uninstall", true);
+            netfilterRegKey.SetValue("UnregDriver", "1", RegistryValueKind.DWord);
+        }
+
+        static bool UninstallStarted()
+        {
+            RegistryKey netfilterRegKey = Registry.LocalMachine.OpenSubKey("SOFTWARE\\Netfilter\\Uninstall", true);
+            if (netfilterRegKey != null)
+            {
+                var unregDriverValue = netfilterRegKey.GetValue("UnregDriver");
+
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
+
+        static void RunRegUtil(RegAction act)
+        {
+            Process pProcess = new Process();
+            pProcess.StartInfo.FileName = "nfregdrv.exe";
+
+            if (act == RegAction.Register)
+            {
+                pProcess.StartInfo.Arguments = Path.GetFileNameWithoutExtension(driverName);
+            }
+            else
+            {
+                pProcess.StartInfo.Arguments = string.Format("-u {0}",
+                    Path.GetFileNameWithoutExtension(driverName));
+            }
+
+            pProcess.StartInfo.UseShellExecute = false;
+            pProcess.StartInfo.CreateNoWindow = true;
+            pProcess.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
+
+            pProcess.Start();
+            pProcess.WaitForExit();
         }
 
         static public bool InstallDriver(ref string errorMsg, DriverType driverType)
@@ -54,7 +115,21 @@ namespace NetfilterInstaller
 
             try
             {
-                File.Copy(srcDriverPath, dstDriverPath);
+                IntPtr oldValue = IntPtr.Zero;
+                if (Wow64DisableWow64FsRedirection(ref oldValue))
+                {
+                    File.Copy(srcDriverPath, dstDriverPath);
+                    if (Wow64RevertWow64FsRedirection(oldValue) == false)
+                    {
+                        errorMsg = "Couldn't revert Wow64 redirection";
+                        return false;
+                    }
+                }
+                else
+                {
+                    errorMsg = "Couldn't disable Wow64 redirection";
+                    return false;
+                }
             }
             catch (FileNotFoundException)
             {
@@ -73,24 +148,69 @@ namespace NetfilterInstaller
             }
 
             //Register driver here..
+            RunRegUtil(RegAction.Register);
 
             return true;
         }
 
-        static public bool DriverAlreadyExits()
+        static public bool DriverAlreadyExits(ref string errorMsg)
         {
             string driverPath = Path.Combine(Environment.ExpandEnvironmentVariables(
                 "%windir%\\system32\\drivers"), driverName);
-            return File.Exists(driverPath);
+
+            bool result = false;
+
+            IntPtr oldValue = IntPtr.Zero;
+            if (Wow64DisableWow64FsRedirection(ref oldValue))
+            {
+                result = File.Exists(driverPath);
+                if (Wow64RevertWow64FsRedirection(oldValue) == false)
+                {
+                    errorMsg = "Couldn't revert Wow64 redirection";
+                    return false;
+                }
+            }
+            else
+            {
+                errorMsg = "Couldn't disable Wow64 redirection";
+                return false;
+            }
+
+            return result;
         }
 
         static public bool DeleteDriver(ref string errorMsg)
         {
             string driverPath = Path.Combine(Environment.ExpandEnvironmentVariables(
                 "%windir%\\system32\\drivers"), driverName);
+
+            IntPtr oldValue = IntPtr.Zero;
             try
             {
-                File.Delete(driverPath);
+                // Unregistrate here..
+                if (!UninstallStarted())
+                {
+                    RunRegUtil(RegAction.Unregister);
+                    SetUninstallKeyStarted();
+                }
+                else
+                {
+                    if (Wow64DisableWow64FsRedirection(ref oldValue))
+                    {
+                        File.Delete(driverPath);
+
+                        if (Wow64RevertWow64FsRedirection(oldValue) == false)
+                        {
+                            errorMsg = "Couldn't revert Wow64 redirection";
+                            return false;
+                        }
+                    }
+                    else
+                    {
+                        errorMsg = "Couldn't disable Wow64 redirection";
+                        return false;
+                    }
+                }
             }
             catch (Exception e)
             {
@@ -98,7 +218,7 @@ namespace NetfilterInstaller
                 return false;
             }
 
-            return !DriverAlreadyExits();
+            return !DriverAlreadyExits(ref errorMsg);
         }
     }
 }
